Initial prototype.
We have entities! Each entity has a transform matrix, a velocity vector, and
an acceleration vector for an uniformly accelerated motion.

So, we have motions! There are a translation, a rotation, and the accelerated motion.
The translation moves the entity according the velocity vector and a time interval.
The rotation rotates the entity according the velocity vector. The accelerated motion
modifies the velocity vector according the time interval. The acceleration vector is
provided only for an information reason.

We could obtain an uniform motion for the entity by setting an initial velocity vector
or by preventing the current acceleration via an oppositely directed acceleration.

Should be mentioned a presentation mechanism! Each entity could have a variable number
of presentations. Presentations are images or visual effects. For example, we have
a transform presentation, which could display the entity position, velocity, and
acceleration.

We work with motions and presentations separetely. So, in the future, we could use
at least two threads for the fast processing.

At the moment, the prototype provides three entities with rectangular presentations:
a background, a dot, and a wall. The background and the wall are static, but the dot
moves with the given acceleration to the right edge of the background.

There is a special entity: a camera. The camera has a transform, and could have a motion.
We use the camera transform in our presentation mechanism. We could setup multiple
cameras for different reasons in the future.