[js] event queue.
We implemented an event processing mechanism. We could set up various actions
using a queue. For example, we use four initial events for the camera, the background,
the dot, and the wall. Each event provides the corresponding entity creation and
behaviour. So, the backgroud and the wall will be displayed static, but the dot
motion will be changed with time. In that way, we set up "onclick" handler for
the canvas after a fixed time interval. This handler enables an interaction with
the dot.